Update and rename dafarkias_Project map navigator.lua to Dfk_Project Map Navigator.lua
I used the ReaPack Upload Tool to create the pull request for this script, and for some reason, the documentation I added was included in the beginning of the script, as code.

Also, I would prefer the titling of the script to be "Dfk_Project Map Navigator," or at very least, "Dfk_Project map  navigator," if for some reason I may not use title case. 

Sorry about the double-change request.